refactor(skills): share modal utilities
