+++ web/src/app/craft/v1/skills/page.tsx
@@ -1,1 +1,1 @@
-export { default } from "@/views/UserSkillsPage";
+export { default } from "@/views/SkillsPage";
